Add files via upload
Implement the Add Order. When the button is clicked, add a new row in the top of the table, with each column showing a text box where data can be entered. Show a button called "Add" under "Action" column. Once all data is entered and "Add" button pressed, the data should be displayed as labelled data in the row.